Initial list of design patterns for these simple recursive-based solutions.
Carol and I were discussing how to approach the various homework problems,
and we came up with this initial list of "design patterns" that we discover
when looking at some of this simple list-processing based recurisions that
we've been working on.

I also added some references to loop-based iterative solutions, since
eventually in our work, we'll start moving towards other programming languages
that tend to use loops rather than recursion to solve these types of
problems.